Design Project/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.10.0
prophet>=1.1
openai>=1.0.0
gradio>=4.0.0
google-colab>=1.0.0
